+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.0.2] - 2024-11-30
+### Added
+- Add reference date and time and write it on the map
+- Implement tight-layout
+- 
 ## [0.0.1] - 2024-11-30
 ### Added
 - Implement temperature interpolation and mapping